Do not output empty MultiLineStrings.
This is similar to a previous change, which attempted to make sure coanacatl doesn't output empty geometries for features. However, it didn't go far enough, and it was possible for a MultiLineString containing only degenerate LineStrings to result in the output of an empty feature with no ID, no properties and no geometry.

Added a test case for this, and tried to move all the `vtzero::*_feature_builder` constructors _after_ any checks to determine whether the geometry is valid and non-empty. In some cases, added new checks to throw exceptions when expected invariants are proved false.